Move course-loading complexity into course-selector
